build: update base operator image to squid
With the stable release of v19.2.1, we update the Rook
operator base image to use this version to ensure
compatibility of ceph commands, in particular for
radosgw-admin.

Images pushed to the ceph quay repo are now manifested
so we can use the --platform flag when building the
docker image instead of pulling from separate repos.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>