build: various Makefile improvements
1. add target checkmake

checkmake can  used for linting the Makefile

2. disentangle messed-up check and test  targets

check was supposed to be an alias for test but the definition was messed up.
In particular, no help text for test was shown.
This fixes the mess by separating the test and check targets.
As a side-effect, the  help text for both is now shown correctly

3.  make target docs phony and add help text

checkmake complained that 'docs'should be phony, so fix that.
Add some help text while at it.

4. reword misleading line to silence checkmake

checkmake was complaining that target 'Options' should be phony.
but Options not a target. It was a line in the Help text.
Fix this by rewording this part of help text

Signed-off-by: Michael Adam <obnox@samba.org>