ci: run selective disk cleanups in go CI tests
The Ceph upgrade tests have been flaky or reliably failing for various
Ceph/Rook versions due to mon out of space issues. Reinstate the disk
cleanup action, but do not clean up items that rely on the package
manager for removal. Package manager takes 2-4 minutes to operate and is
likely not worth it, even for saving 15+ GB of space.

From testing with this configuration, we can save 11GB of space in about
8 seconds of runtime. This should (hopefully) be plenty of space to keep
total disk usage below 95% and keep mons healthy, and the time needed is
negligible.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>