docs: slightly improve wording in release guide
Reading the release process documentation, I found the description
of steps in one place to not quite match the intended effect of the
actual command steps due to a slight imprecision in used terminology.

This change improves and clarifies the corresponding wording to hopefully
prevent others from facing the same confusion.

Signed-off-by: Michael Adam <obnox@samba.org>