osd: Fix removal when cluster in different namespace
The `rook ceph osd remove` command would try to delete Deployments and
PersistentVolumeClaims in the same namespace as the Rook operator, which
would fail with

```
cephosd: failed to fetch the deployment "rook-ceph-osd-0". deployments.apps "rook-ceph-osd-0" not found
```

if the Ceph cluster is deployed in a different namespace.

This commit makes `rook ceph osd remove` look for Deployments and
PersistentVolumeClaims in the namespace named after the cluster itself,
which is the standard pattern.

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>