ci: increase logging and time for OBC lifecycle rm test
Nightly CI has been failing recently having issues with removing a
bucket's lifecycle config after initially applying it. A Ceph RGW issue
is suspected, but more information is needed. Attempt a longer timeout,
to see if it is a slow async process. Also log the returned info to see
if it is helpful.

This commit is intended to be eventually reverted, but it must be
checked into master so that nightly CI runs provide the needed info.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>