ci: only publish docs in master and tagged releases
Until now, docs have been published for every master and
release branch build. This can lead to confusion if a new
feature is backported to the release branch and the docs
are published to the website, even though the feature will
not be available until the next tagged release.

This does have the effect that small doc-only fixes will
not be in the latest release docs until we actually tag
a release.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 28fa387dff189b5b8ec949bb136538229bc24d29)