Don't use user secret with admin secret in import-external-cluster.sh
When connecting to an external cluster, currently you can run the script with only `ROOK_EXTERNAL_ADMIN_SECRET` set.  However, currently it'll use the empty `ROOK_EXTERNAL_USER_SECRET` when creating the secret. This causes the operator to fail with `the cluster identity was not established: ceph secret is empty`. After this change, you can successfully connect to the cluster.