doc: fixup bullet points on rgw docs
In #15961 I updated the rgw docs, but I was looking at the published
docs and the bulleting is broken now for Gateway Settings. When I was
working on the PR I was just using the markdown preview in my editor and
wan't using the `docs-preview` target. It turns out that the markdown
editor and mkdocs render things differently.

This PR fixes the bullet formatting and I verified the fix by running
`make docs-preview` on my box and double checking the local website.

Signed-off-by: Elias Carter <elias@dropbox.com>