namespace: log right deletion msg in rns
With the current code, we are never logging with a deletion msg when
there is no image inside RNS. We have a code check for `if containsImages`
and inside that, the first check we have is for `if containsImages`.
It will never log `else` condition, where we logs.

```
rados namespace <RNS_Name> is empty and can be deleted
```

Additionally, for blockpool, before returning an error during blockpool deletion, we want to update the status in CR. Therefore, we are moving the error return statement after the CR is updated with a log.

Signed-off-by: subhamkrai <srai@redhat.com>