core: Only support isa and jerasure algorithms on EC pools
The `CephBlockPool` CR allowed any algorithm to be specified in
`spec.erasureCoded.algorithm`, but in practice only `isa` or `jerasure`
would work.

Make that restriction explicit by restricting this field to `isa` or
`jerasure` in the CRD.

Also make the `ErasureCoded` spec parameters immutable, as it's not
possible to modify the erasure-code profile after the pool has been
created in Ceph.

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>