helm: refactor to reduce repetition, use with instead of if
A `with` block won't render if the argument is falsy, so it behaves like
an `if` block that redefines the `.` context. Using `with` instead of
`if` can help reduce repetition, making things more readable in the
process.

Signed-off-by: Erik Sundell <erik@sundellopensource.se>
(cherry picked from commit 8f60b33f314b3ac0b7efb1290f0581d04c02ff8e)