ci: osd with metadata partition device running incorrect ceph version
The osd-with-metadata-partition-device canary test was always
running the version of ceph found in cluster-test.yaml,
instead of the intended version set for the canary tests to
run with in the ceph_images ci variable.
Thus we were seeing this test be more unstable when running
against ceph Tentacle.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 3ca04322466dbcc490f03ef72433faf94532eb4d)